Delete account

+++ app/src/main/java/com/example/chefhub/ui/AppViewModel.kt
@@ -1,6 +1,7 @@
 package com.example.chefhub.ui
 
 import android.content.Context
+import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.chefhub.db.ChefhubDB
@@ -10,6 +11,7 @@ import com.example.chefhub.db.data.Follows
 import com.example.chefhub.db.data.Recipes
 import com.example.chefhub.db.data.Users
 import com.example.chefhub.db.repository.RecipesRepository
+import com.example.chefhub.screens.components.clearCredentials
 import com.example.chefhub.screens.components.showMessage
 import com.google.firebase.Firebase
 import com.google.firebase.auth.FirebaseAuth
@@ -466,29 +468,39 @@ class AppViewModel(private val database: ChefhubDB): ViewModel() {
     }
 
     fun deleteAccount() {
-        // COMENTARIO.
         val auth: FirebaseAuth = Firebase.auth
         val user = appUiState.value.user
 
-        // COMENTARIO.
-        viewModelScope.launch {
-            auth.signOut()
-            database.usersDao.deleteUser(user)
+        try {
+            auth.currentUser?.delete()?.addOnCompleteListener { task ->
+                if (task.isSuccessful) {
+                    viewModelScope.launch {
+                        try {
+                            database.usersDao.deleteUser(user)
 
-            _appUiState.update { currentState ->
-                currentState.copy(
-                    user = Users(),
-                    viewedUser = Users(),
-                    users = arrayListOf(),
-                    followers = arrayListOf(),
-                    following = arrayListOf(),
-                    recipe = Recipes(),
-                    recipes = arrayListOf(),
-                    favorites = arrayListOf(),
-                )
+                            _appUiState.update { currentState ->
+                                currentState.copy(
+                                    user = Users(),
+                                    viewedUser = Users(),
+                                    users = arrayListOf(),
+                                    followers = arrayListOf(),
+                                    following = arrayListOf(),
+                                    recipe = Recipes(),
+                                    recipes = arrayListOf(),
+                                    favorites = arrayListOf(),
+                                )
+                            }
+                        } catch (dbException: Exception) {
+                            println("Error.")
+                        }
+                    }
+                }
             }
+        } catch (exception: Exception) {
+            println("Error.")
         }
     }
+
 
     /** Funciones Database **/
     fun onSelectUser(user: Users) {
